+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `Toggletip` component.
+
 ## [2.119.0] - 2020-02-10
 
 Feature: